backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pypdf
python-docx
python-dotenv
